Refactored the codebase
